Modified Structure, Add Tests, Fix Issue
