+++ package.json
@@ -1,11 +1,13 @@
 {
     "name": "strong-password-check",
-    "version": "1.0.1",
+    "version": "1.0.2",
     "description": "Check the strength of the password inputed and returns what needs to be updated in the password for that to be more strong.",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
+        "test": "jest --config jestconfig.json",
         "build": "tsc",
-        "prepublishOnly": "npm run build"
+        "prepublishOnly": "npm run build",
+        "format": "prettier --write \"src/**/*.ts\" \"src\"",
+        "lint": "tslint -p tsconfig.json"
     },
     "repository": {
         "type": "git",
@@ -68,5 +70,13 @@
     "bugs": {
         "url": "https://github.com/nihalavulan/strong-password-check/issues"
     },
-    "homepage": "https://github.com/nihalavulan/strong-password-check#readme"
+    "homepage": "https://github.com/nihalavulan/strong-password-check#readme",
+    "devDependencies": {
+        "@types/jest": "^29.5.2",
+        "jest": "^29.5.0",
+        "prettier": "^2.8.8",
+        "ts-jest": "^29.1.0",
+        "tslint": "^6.1.3",
+        "tslint-config-prettier": "^1.18.0"
+    }
 }
