--- src/constants/defaultConfig.ts
@@ -1,11 +1,0 @@
-import { DefaultConfig } from "../types";
-
-const defaultConfig: DefaultConfig = {
-    lowercase: true,
-    uppercase: true,
-    digits: true,
-    specialChars: true,
-    minLength: 8,
-};
-
-export default defaultConfig;
